refactor(editor): extract placement tray commands
Move canvas drops, deterministic Place all, display-annotation creation, and return-to-tray semantics into a component-placement controller. App now supplies document state, canvas coordinate projection, selection, and transaction boundaries.

Validation: pnpm gate:plan -- --path apps/editor/src/features/component-insert/placement-tray-commands.ts --path apps/editor/src/app/App.tsx; pnpm typecheck; pnpm test:local apps/editor/src/features/component-insert/placement-tray.test.ts apps/editor/src/features/component-insert/placement-tray-panel.test.tsx apps/editor/src/app/App.test.tsx; pnpm test:e2e:local apps/editor/e2e/component-insert.spec.ts --grep returns a component to the Placement Tray and places the retained Instance again; pnpm ci:static; git diff --check

Test-Impact: no-test-change — existing placement planning, App, and tray browser contract cover unchanged drop/place/return behavior
